Remove unused applyUserOptions
Removing the unused implementation of applyUserOptions.
applyUserOptions was implemented only on Power so that on startup it was possible to disable P10 support when TR_EnableExperimentalPower10Support is not set.
applyUserOptions was ultimately implemented but never used.
Since the method is not implemented on any other platform, it should be removed from the class.

Signed-off-by: AlenBadel <Alen.Badel@ibm.com>