Fixup JitBuilder tracing output
Under traceIlGen, JitBuilder provides more output on how it is
being used to generate OMR compiler IL. In particular, it logs
every operation as it is called. Unfortunately, some operations
are missing this output, some of the output is missing key
operands (for example, Calls did not list their operands or
return value), and some of the output was inconsistently
formatted (return values weren't always shown in the same
way). This commit fixes all these problems.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>