Fix handling of non-volatile unresolved symrefs on ppc32
As it turns out, under certain corner cases we can generate unresolved
symrefs that are not marked as volatile. In the event that such a symref
is loaded as a long on ppc32, LoadStoreHandler would get confused and
fail to use its handling for unresolved on such symrefs. Instead, it
would generate two separate loads/stores, leaving an unused snippet that
will cause a crash later. This has been corrected and LoadStoreHandler
now checks for unresolved before checking volatility.

Signed-off-by: Ben Thomas <ben@benthomas.ca>