Add flag to disable loadaddr specialization in x86 MemoryReferences
Presently in the x86 code generator, there is an optimization during
MemoryReference creation that always assumes that accesses to local
objects via a loadaddr should use the frame register, as opposed
to evaluating (or re-using) the register associated with the loadaddr
as the base register.

Add an environment variable `TR_useLoadAddrRegisterForLocalObjectMemRef`
to disable this optimization for diagnostic purposes.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>