Initialize TR_X86ProcessorInfo once
TR_X86ProcessorInfo is currently used in OMR because the compiler does
not yet have access to the portlib. Therefore, it has to determine the
processor features itself. In order to not initialize
TR_X86ProcessorInfo multiple times a flag is set. However, it is
a race possible if there are multiple compilations, as is the case in
Eclipse OpenJ9. Because the processor feature flags were copied from the
target TR::Environment set in TR::Compilation object, if two racing
compilations had different target environments, the feature flags could
become inconsistent.

This commit initializes the processor info just once in
OMR::CompilerEnv::initializeTargetEnvironment().

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>