Define isArrayCompTypeValueType method for Value Propagation
This change adds a virtual method and base implementation of
isArrayCompTypeValueType to OMR::ValuePropagation that can be used to
determine whether the component type of an array is definitely, might be
or is definitely not a value type.

The base implementation simply assumes any array's component type might
be a value type if value type support is enabled, and that it is
definitely not a value type otherwise.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>