Reorg Scavenger::copy() epilog
In copy method we checked 2x if forwarding succeeded. It's needed for
Concurrent Scavenger for 2 different paths: small objects
(allowing duplicates) and large objects (disallowing duplicates).

These 2 if checks (and their associated then&else code blocks) were
serialized, but now they are nested. It makes it somewhat easier to
read, but more importantly more friendly for later code versioning
for CS and nonCS case.

A small compromise is that handling of failed forwarding path now is
duplicated at the level of source code (now at each end of those 'if's,
while it used to be only once at the very end). Still, in run time, it's
only handled once.

Signed-off-by: Aleksandar Micic amicic@ca.ibm.com