Fix paddi and lxvdsx for vsplatsEvaluator to have GPR register
P10 uses paadi in conjuction with lxvdsx for vsplats. The opcode
paddi requires a GPR for correct address calculation as lxvdsx
uses GPR for loading the data. The passed register for
loadfloatConstant can be of type FPR/VSX thus leading to assert
in RegisterAssigning phase. This commit ensures that paddi and
lxvdsx uses GPR for the constant data snippet address.

Fixes: openj9#12383

Signed-off-by: Mohammad Nazmul Alam <mohammad.nazmul.alam@ibm.com>