Deprecate findOrCreateLiteral codegen API
This API is unimplemented and calling it and using the result will
likely produce a crash. This API is used in dead code paths:

1. DFP
2. arraycmpWithPad
3. pdneg
4. vsetlem

under certain circumstances. Please see the self review in the PR which
is associated to this commit for explicit reasons as to why failing the
compilation in each case is safe and warranted.