CMake: update z/OS export handling
Copy the .x files to the ARCHIVE_OUTPUT_DIRECTORY to be consistent with the
CMake model. However, we continue to copy the .x files to the
RUNTIME_OUTPUT_DIRECTORY as well, to work around a bug in CMake which causes
it to look for the .x files in the wrong directory.

Signed-off-by: Devin Nakamura <devinn@ca.ibm.com>