Deprecate dead APIs used in doRegisterAssignment
The APIs for counting various statistics has been disabled by default
since it is guarded by the `DEBUG` build time flag which we don't build
with. This code is dead. Once we are further along unifying the register
allocator across codegens we can come back and re-introduce these stats
in a cross-platfrom way.