Unify TR_*OutOfLineCodeSection::assignRegisters across codegens
This API is almost identical on all our codegens. There are very small
differences which have been `#ifdef` out. We're currently not 100% sure
that these guarded paths are needed so we will leave them in for now
and look to eventually unify the entire function as we get closer to
true codegen unification.