Enable OSR for profiling compilations with JProfiling
In the profiling compilations that was using JIT Profiler, we had to
disable the OSR in profiling compilations as we had to duplicate the
method and size would become an issue and profiling compilations are
very short-lived. With JProfiler we do not need to duplicate the method
body and counters to collect profiling data are inserted in the code.
This commits enables OSR and sets HCR mode to osr from traditional for
profiling compilations when JProfiler is used which would allow us to
collect accurate profile of the method to be used in subsequent
compilations which by default has OSR enabled.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>