Add command line option to make inliner more aggressive
Added `enableAggressiveInlining` command line option which, under the covers,
will enable other options or set threshold values such that the Inliner is
more aggressive in inlining callees. This is done in the `jitPostProcess()`
phase, so it may override other inliner specific options, also present on
the command line.
The new option is directed at experimentation and not geared towards
production use. Moreover, the new option cannot be applied in a subset,
but only globally, to all the methods.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>