Allow inliner policy to decide whether to remove differing targets
Previously, for direct calls, targets that appear to be a different
method from the one specified by the call node's symbol reference would
always be removed.

This behaviour is still the default, but the new query allows downstream
projects to customize the conditions under which those targets are
removed, e.g. by exempting certain methods.