Use addAtomic for incrementing counters of OMRMemCategory
Use addAtomic for incrementing counters of OMRMemCategory instead of
spinning with compareAndSwapUDATA.
Similarly, use subtractAtomic for decrementing counters.

Because this change simply makes omrmemcategories.c to use existing
implementation of atomic add/sub in AtomicSupport.hpp and they are identical
to replaced implementation in omrmemcategories.c, this commit should introduce
no functional or performance change, but this will enable us to optimize
implementation of atomic add/sub in a single file in the future.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>