Print OMR options in effect into the verbose log
Options that don't have the first character of the message as "F" in the option table entry will not
be printed in "options in effect" when the option is specified.
Thus, I added 'F' to the messages of OMR options, so that the options specified would also
be printed to "options in effect" in verbose logs.

similar to the PR in openj9: #12790

Signed-off-by: Eman Elsabban <eman.elsaban1@gmail.com>