Migrate warnings as error queries outside of create_omr_compiler_library
Downstream projects may want to override warnings as errors for compiler
libraries. Specifically, when compiling NASM sources on Windows Cygwin
we want to avoid adding the warnings as error options via the
`target_compile_options` API because NASM on Cygwin does not accept the
`/WX` flag which Visual Studio wants to apply. In such cases the
assembly sources are separated into their own compiler library and
warnings as errors are disabled before creating the library such that
the aforementioned flag is not applied.