Apply warnings flags only to C/CXX sources
The flag is not compatible with all sources within a target. When we mix
assembly and C/C++ sources the warnings as error flag will be applied to
assembly sources as well, and the assember used may not support this
flag. Because the flag is C/C++ compiler specific at the moment, we only
add the flag for those targets.