Add DummyResolvedMethod flag to OMR Symbol
Includes getter/setter functions. This flag is used to identify
symbols corresponding to dummy TR_ResolvedMethod that are not
really resolved, but treated as such as the resolution mechanism
involves resolving the arguments instead. An example of that is
linkToStatic, which is a VM internal native call that is created
for unresolved invokedynamic and invokehandle bytecodes.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>