Fix createIntRangeConstraint warning
In a previous commit we erroneously changed the constraing for some VP
handlers to address warnings in the area. We are isolating the this fix
because the previous chnages were functionally incorrect. We address the
warning instead with a template overload for `createIntRangeConstraint`.

See the following review comment for further discussion:
https://github.com/eclipse/omr/pull/6030#discussion_r641864997