Get real MacOS product version if in compability mode
If a program is built by a 10.15 or earlier SDK is run on Big Sur (version 11)
or newer, it will run in a backwards compatibility context which will return
product version as 10.16.

Thus, when we need to get Mac OS version when running on 11+, we need to use
a workaround to get the real version string. In this case, we look at a symlink
to the SystemVersion.plist file which bypasses the backwards compatibility
context.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>