Refactor copy to use template for concurrent GC
Copy is now an inline function which calls copyForVariant, which
performs the previous functionality for copy. copyForVariant is a
template function that is told the variant of template that it is to
use, either CS or STW. This dictates if copyForVariant should
exhibit Concurrent Scavenger behaviour or Stop The World behaviour.

The enum CopyVariant is introduced to represent the different variants
of copyForVariant. The two options are STW for IS_CONCURRENT_ENABLED
== false, and CS for IS_CONCURRENT_ENABLED == true.

Signed-off-by: Oussama Saoudi <oussama.saoudi@ibm.com>