Refactor opcode parametarizations in preparation for migration
To migrathe some of the parametarized opcodes we eliminate the template
definitions if favour of macros which we will move into the `InstOpcode`
class in the following commit. This change should not have any effect on
semantics of the code and it is purely stylistic.

The previous `#define` definitions of these mnemonics would not work
because of the C preprocessor, so we could not properly qualify the
mnemonics with `TR::InstOpCode::` in code because the function
definitions of the parameterized mnemonics do not exist in that class.
We have to have actual functions there for the name resolution to work,
so we had to move away from the `#define` to an inline function.