Add compilePortableCode query
Add a new query to the TR::Compilation class to allow the compiler to
query whether the runtime is some kind of mode (such as
snapshot/restore) that would necessitate the compiler to generate
code that is portable (ie generated on one system but can run on
another). This is different from relocatable compiles because, for
example with snapshot/restore, the code does not need to be relocated
but it does need to be portable.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>