Assert that labels are defined when applying label relocations
A label has a null address until its location is determined when it is
encountered in the instruction stream during binary encoding. When the
compiler generates a jump to a label whose address has not yet been
determined, it also generates a label relocation to correct the jump
target once the label's address is known after binary encoding. But if
for some reason the label's address is never updated, the relocation
would silently continue using zero and generate an incorrect jump.

Because such jumps can be hard to debug, this commit adds assertions
that verify that the label has at least had some address set, otherwise
failing fast. Note that zero cannot be the true address of a label,
since we never generate instructions there.