Improve registerExchange API on Z
When assigning virtual register to target real register while assigning
registers for instruction, register exchange API which would take care
of care of making exchange between the target real register and current
assigned register bails out when faced with different sized register and
falls back to the path where we would need spare register to satisfy the
requirement. This would cause an issue where we run out of all the
registers.
In order to make exchange between target real register and current
assigned register, if either one is 64-bit, we can simply use 8 byte
spill slot to do that. This commit upgrades the register exchange API to
do that.

Fixes: https://github.com/eclipse-openj9/openj9/issues/12789

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>