Modified Concurrent Scavenger conditions to use compiler optimization
Rason for condition change:
- allowDuplicate bool only set to true if CS is enabled. Compiler
can skip condition check for allowDuplicate if
OMR_GC_CONCURRENT_SCAVENGER is enabled but CS is disabled.
- nested if-forwarding-succeeded check succeeds in forwarding
the object if CS is disabled. This is for the case
OMR_GC_CONCURRENT_SCAVENGER but CS is disabled. Compiler can
go directly into succeeded logic
- if OMR_GC_CONCURRENT_SCAVENGER enabled but CS is disabled,
there's no need to check allowDuplicateOrConcurrentDisabled,
so we skip the block.

Signed-off-by: Oussama Saoudi <oussama.saoudi@ibm.com>