Treat warnings as errors on z/OS xlC v2.2
The z/OS xlC v2.2 compiler does not support the `-Werror` option that is
available starting with the v2.3 compiler version. Instead it does
support the `-qhalt=4` option which according to the documentation:

> Stops compilation process of a set of source code before producing any
> object, executable, or assembler source ®les if the maximum severity
> of compile-time messages equals or exceeds the severity specified for
> this option.

Adding this option to the command line unfortunately does not achieve
the desired effect of enabling warnings as errors since compilations
appear to complete with rc=0. The reason for this is because `HALT` is
a batch option, not a USS one. To get the desired effect we have to
specify two environment variables which when active will treat warnings
as errors and return a non-zero return code. The two options are:

```
export _C89_ACCEPTABLE_RC=0
export _CXX_ACCEPTABLE_RC=0
```

for C and C++ respectively. Until we can upgrade to the v2.3 compiler we
have to run with these environment variables active to get the desired
effect.