Fix CCN3296
If we are building using the native codepage (EBCDIC) we do not need to
include this header. Otherwise `OMR_EBCDIC` will be defined and we will
be linking with the atoe library so this header will be on the include
path and we can safely include it.

```
WARNING CCN3296 /openzdk/jenkins/workspace/Build/port/unix_include/omriconvhelpers.h:38    #include file "atoe.h" not found.
```