Rename LightweightNonReentrantReaderWriterLock to avoid HLASM errors
If the file name of the compilation unit is greater than 38 characters
and the compilation unit uses inline assembly HLASM will fail with the
following message:

```
INFORMATIONAL CCN1149: Inline assembly statements caused HLASM RC=4 with the following messages:
WARNING CCN1150: Inline assembly statements caused HLASM RC=4.
```

The listing file says the assembler has succeeded so this error message
is quite puzzling and seems to go away when the file name is shortened.
It is likely the CSECT name is too long for HLASM to handle, although
changing the CSECT name does not alleviate the problem.

The simplest solution seems to be to truncate the name a little bit
which is what we've done here.