Fix IEW2646W
Here is the full warning message:

```
IEW2646W 5383 ESD RMODE(24) CONFLICTS WITH USER-SPECIFIED RMODE(ANY) FOR
          SECTION $PRIV00007C CLASS B_TEXT.
FSUM3065 The LINKEDIT step ended with return code 4.
```

Taking command-line for the lnik step of an executable which prints this
message, ex. omrporttest, we can add the `-Wl,MAP` option to print some
debugging diagnostics. From here we can see the offending section:

```
 ---------------
 CLASS  B_TEXT            LENGTH =     2BAC  ATTRIBUTES = CAT,   LOAD, RMODE= 24
                          OFFSET =        0 IN SEGMENT 002     ALIGN = DBLWORD
 ---------------

  SECTION    CLASS					------- SOURCE --------
   OFFSET   OFFSET  NAME		TYPE    LENGTH  DDNAME   SEQ  MEMBER
...
...
              21A0  $PRIV00007C     *  CSECT        10  /000001D  01  omrgetstf-64.s.o
	0     21A0     getstfle           LABEL
```

Looking at the contents of this file though we note that the actual
CSECT starts halfway through the file. The fact that the section was
named "$PRIV00007C" must mean that it is referring to the code before
the CSECT. Hence the way to fix this problem is to define the CSECT
right at the start of the file and to define RMODE as ANY since it
defaults to RMODE 24 if it is not specified as per [1].

[1] https://www.ibm.com/docs/en/zos/2.1.0?topic=binder-amode-rmode-attributes