Cache known object info on TR::Nodes
Allow known object information to be associated with arbitrary nodes.
This will permit more accurate association of refined known object
info, and allow it to be placed on arbitrary nodes.

A new 32-bit bit field is added to the Node class in the padding
space between two fields.  This incurs no TR::Node footprint increase
on 64-bit architectures.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>