Define symbol for new jitAcmpneHelper
The existing jitAcmpHelper symbol can be used in downstream projects to
represent object reference comparisons for equality.  This change
introduces a jitAcmpneHelper symbol to represent inequality comparisons
rather than treating that as a second-class operation requiring negation
to implement.

The change also renames jitAcmpHelper to jitAcmpeqHelper to clarify the
purpose of the existing symbol.  The old symbol name will be retained
until it is no longer being used by downstream projects.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>