(0.28.0) AArch64: Correctly handle ificmpeq/ificmpne nodes with many global registers
Correctly count the integer or address child nodes added to GlRegDeps node
to avoid adding too many integer registers to the register dependencies.
Add a src register of compare branch instructions to register dependencies
so that a src register is assigned when other registers in the register dependencies
are assigned.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>