Restrict rotate instruction optimization for land nodes
This commit limits an optimization in the z codegen that tries
to perform a shift + land operation using a single rotate instruction.
The optimization skips the evaluation of the land node but evaluates its
children. If the reference count of the land node > 1, then this can
result in the code generator double decrementing the land node's children.
This commit restricts this optimization to when the land node's ref count
is equal to 1 so that the above described scenario is no longer possible.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>