Define TR_jitLookupDynamic[Public]InterfaceMethod JIT helpers
These new JIT helpers are OpenJ9-specific but unfortunately they still
need to be declared here.

The definition of TR_jitLookupDynamicPublicInterfaceMethod uses
TR_newArray+2 instead of TR_jitLookupDynamicInterfaceMethod+1 because
the latter causes s390m4check.pl to complain that there is a "line
longer than 71 characters". It's difficult to add this comment inline
next to the definition itself because Helpers.inc is included directly
into all of the following file types:
- C++ header
- CPP-preprocessed assembly
- m4-preprocessed assembly