Remove SupportedForPRE from instanceof for Field Privatizer bug
instanceof operations can be hoisted out of loops by PRE.  This could
expose a bug in Field Privatizer in which a field access that is guarded
might be privatized, resulting in the field being referenced and stored
to unconditionally.

A temporary patch in Field Privatizer checks for uses of instanceof and
null comparisons inside of loops to guard against the conditions that
expose the bug.  This change removes the SupportedForPRE property from
the instanceof operator to prevent its being hoisted out a loop,
avoiding one more situation that could expose the bug.

Once the bug is fixed properly in Field Privatizer, the SupportedForPRE
property will be added back to instanceof.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>