Fix core file not generating on OSX
When reading memory segments, reset the return value kr so that the
generated file is not deleted at the end if the last memory segment
throws an error when being read, since mach_vm_read errors are not
fatal.

Also make sure final return value is correct when there is an error.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>