Split _options and _optionString in OptionSet into separate fields
When both AOT and JIT have logs specified and when one is processed
first, the OptionSet._options for the other one has not been set up
properly yet. Since setOptions has not been called for the other
command option, getOptions returns a char pointer to _optionString
that is implicitly casted as TR::Options * because OptionSet._options
and OptionSet._optionString are union. To properly check if
OptionSet._options is valid, it needs to be a separate field.

Fixes #6189

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>