Improve native stack traces on Linux
Using dl_iterate_phdr(3) we can find the loaded program element
containing a given address. When the absolute file name is
available, we can then open the ELF shared library or executable
to find the symbol table section and related string table section
to derive a name and offset for that address.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>