Simplify load of J9Class flags off fix loadaddr
Value propagation optimization which simplifies loading of class and
flags off J9Class pointer of fixed type expects the child of the
indirect load to be a VFTLoad from an object. Other value propagation
transformation might optimize the VFT load from an object of fixed and
known type and convert it to constant node to load constant address.
This commit updates the simplifyJ9ClassFlags to transform load of
J9Class flags and pointer of a fixed known class to constant load.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>