Fix FieldPrivatizer so that it does not miss emitting write barriers
The Field Privatizer optimization needs to write back privatized field
stores on loop exits. This fix will ensure that write barrier field
stores will be used as the basis of the exit store back node to
ensure that loop exit stores do not skip write barrier operations when
some original non-privatized store inside the loop was doing the
operation. Previously a field load could have been used an the basis of
the exist loop store resulting in an astore rather then an awrtbar.
https://github.com/eclipse/omr/pull/6321

Signed-off-by: Kevin Langman <langman@ca.ibm.com>