SATB TLH Batch Mark
Objects allocated while CM (SATB) is active must be marked to preserve
the invariant required to guarantee correctness.

Instead of marking each obj individually, we can mark on the TLH level.
However, doing so is non-trivial: we must know the start of the last
object to mark the proper range of the TLH. That is, arbitrary bits at
the end of the TLH can't be marked, and this is problematic for sweep.

To get around this issue we seal the TLH with a bogus/dummy obj at the
end of the TLH. With this, we know the precise addr to mark up to.
However, we must guarantee that we have at least min obj size available
in the TLH.

These changes introduce a mechanism for the collector to reserve bytes
in a TLH. Whenever the TLH is refreshed, it queries the collector to
determine if it should reserve space for it. If so, part of TOP of the
TLH is hidden. Once the TLH is about to be flushed/cleared, we restore
the the TLH Top and notify the collector.

Signed-off-by: Salman Rana <salman.rana@ibm.com>